% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_mbr_app_helpers.R
\name{mbr_app_metric_choices}
\alias{mbr_app_metric_choices}
\title{Get the metric choices for the current setting}
\usage{
mbr_app_metric_choices(setting)
}
\arguments{
\item{setting}{Character string.}
}
\value{
A named character vector.
}
\description{
Get the metric choices for the current setting
}
\keyword{internal}
