% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_mbr_app_helpers.R
\name{mbr_app_run_code}
\alias{mbr_app_run_code}
\title{Generate a simple run code}
\usage{
mbr_app_run_code()
}
\value{
A character run code.
}
\description{
Generate a simple run code
}
\keyword{internal}
