% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_mbr_app_helpers.R
\name{mbr_app_run_simulation}
\alias{mbr_app_run_simulation}
\title{Run the app-selected simulation}
\usage{
mbr_app_run_simulation(input)
}
\arguments{
\item{input}{A list-like object containing app inputs.}
}
\value{
A data frame of simulation results.
}
\description{
Run the app-selected simulation
}
\keyword{internal}
