% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_mbr_app_helpers.R
\name{mbr_app_seed}
\alias{mbr_app_seed}
\title{Convert an optional seed input to integer}
\usage{
mbr_app_seed(seed_input = "")
}
\arguments{
\item{seed_input}{Character string.}
}
\value{
An integer seed or \code{NULL}.
}
\description{
Convert an optional seed input to integer
}
\keyword{internal}
