% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_mbr_app_helpers.R
\name{mbr_app_text_to_vector}
\alias{mbr_app_text_to_vector}
\title{Convert text input to a vector}
\usage{
mbr_app_text_to_vector(text_input)
}
\arguments{
\item{text_input}{Character string.}
}
\value{
A vector parsed from the text input.
}
\description{
Convert text input to a vector
}
\keyword{internal}
