% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_mbr_app_helpers.R
\name{mbr_append_input_params}
\alias{mbr_append_input_params}
\title{Append app input parameters to simulation results}
\usage{
mbr_append_input_params(df, input)
}
\arguments{
\item{df}{A data frame of simulation results.}

\item{input}{A list-like object containing app inputs.}
}
\value{
A data frame with appended app parameters.
}
\description{
Append app input parameters to simulation results
}
\keyword{internal}
