% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr_helpers.R
\name{mbr_check_paper_metric_inputs}
\alias{mbr_check_paper_metric_inputs}
\title{Check that paper-style metrics are being used in a two-arm 1:1 setting}
\usage{
mbr_check_paper_metric_inputs(ratio, labels)
}
\arguments{
\item{ratio}{Integer allocation ratio.}

\item{labels}{Character labels for each treatment arm.}
}
\value{
Invisibly returns TRUE.
}
\description{
Check that paper-style metrics are being used in a two-arm 1:1 setting
}
\keyword{internal}
