% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr_helpers.R
\name{mbr_complete_randomisation}
\alias{mbr_complete_randomisation}
\title{Generate a complete randomisation allocation}
\usage{
mbr_complete_randomisation(n, ratio = c(1, 1), labels = NULL)
}
\arguments{
\item{n}{Final allocation length.}

\item{ratio}{Integer allocation ratio.}

\item{labels}{Character labels for each treatment arm.}
}
\value{
A character vector of length \code{n}.
}
\description{
Generate a complete randomisation allocation
}
\keyword{internal}
