% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr_helpers.R
\name{mbr_multi_allocation}
\alias{mbr_multi_allocation}
\title{Generate merged block randomisation allocations for multiple centres}
\usage{
mbr_multi_allocation(n, ratio = c(1, 1), labels = NULL)
}
\arguments{
\item{n}{Integer vector of allocation lengths.}

\item{ratio}{Integer allocation ratio.}

\item{labels}{Character labels for each treatment arm.}
}
\value{
A list with one allocation vector per centre.
}
\description{
Generate merged block randomisation allocations for multiple centres
}
\keyword{internal}
