% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr_helpers.R
\name{mbr_pbr_allocation}
\alias{mbr_pbr_allocation}
\title{Generate a permuted block randomisation allocation}
\usage{
mbr_pbr_allocation(n, ratio = c(1, 1), labels = NULL)
}
\arguments{
\item{n}{Final allocation length.}

\item{ratio}{Integer allocation ratio.}

\item{labels}{Character labels for each treatment arm.}
}
\value{
A character vector of length \code{n}.
}
\description{
Generate a permuted block randomisation allocation
}
\keyword{internal}
