% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr_helpers.R
\name{mbr_permuted_block}
\alias{mbr_permuted_block}
\title{Build one permuted block}
\usage{
mbr_permuted_block(ratio, labels)
}
\arguments{
\item{ratio}{Integer allocation ratio.}

\item{labels}{Character labels for each treatment arm.}
}
\value{
A character vector containing one randomly permuted block.
}
\description{
Build one permuted block
}
\keyword{internal}
