% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr_helpers.R
\name{mbr_suballocation_imbalance_prop}
\alias{mbr_suballocation_imbalance_prop}
\title{Compute the proportion of prefixes at or above an imbalance threshold}
\usage{
mbr_suballocation_imbalance_prop(
  allocation,
  imbalance_threshold = 2,
  labels = NULL
)
}
\arguments{
\item{allocation}{Character allocation vector.}

\item{imbalance_threshold}{Imbalance threshold.}

\item{labels}{Character labels for the two treatment arms.}
}
\value{
Numeric scalar.
}
\description{
Compute the proportion of prefixes at or above an imbalance threshold
}
\keyword{internal}
