% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr_helpers.R
\name{mbr_validate_ratio_labels}
\alias{mbr_validate_ratio_labels}
\title{Validate ratio and labels for allocation generators}
\usage{
mbr_validate_ratio_labels(ratio, labels = NULL)
}
\arguments{
\item{ratio}{Integer allocation ratio.}

\item{labels}{Character labels for each treatment arm.}
}
\value{
A list containing validated \code{ratio} and \code{labels}.
}
\description{
Validate ratio and labels for allocation generators
}
\keyword{internal}
